Outsource ThemeToggleButton from Navigation to component
